refactor: scope agent mention serialization to webui
Keep CLI agent references as plain slash text and colocate structured mention serialization with its WebUI consumers.
