+++ packages/common/package.json
@@ -11,7 +11,6 @@
     "./device-link": "./src/device-link/index.ts",
     "./device-link/client": "./src/device-link/client.ts",
     "./message-utils": "./src/message-utils/index.ts",
-    "./agent-mention": "./src/agent-mention.ts",
     "./git-utils": "./src/git-utils.ts",
     "./diff-utils": "./src/diff-utils.ts",
     "./env-utils": "./src/env-utils.ts",
