+++ packages/vscode-webui/src/lib/serialize-custom-agent-mention.ts
